move create_radiation_field() function
